Add files via upload
A simple example of javascript:
As any webapplication works there is a client side and server side to a it. As django works on the server side, javascript is the code on the client side which makes it very interactive for the user. It has no relation with the server side of the code. 
<script></script> - the code return in between these tags can be considered as javascript code.
The simple alert function shows an alert as soon as the page is loaded.