Add files via upload
It is an example of single page application. As it is long time to load different pages for every view, single page application comes handy. There might be certain sections that are common to all pages which need not be loaded everytime. It will be helpful if we just remove certain things on page and add new things to the same page for faster loading of the page.